# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gateway - Food Delivery Microservices",
    "version": "1.0.0",
    "description": "Dokumentasi OpenAPI terpusat untuk semua layanan di belakang API Gateway."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Gateway Lokal"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT yang diperoleh dari endpoint /auth/login."
      }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "restaurant_id": { "type": "integer" },
          "total_price": { "type": "number", "format": "float" },
          "status": { "type": "string", "enum": ["PENDING", "PAID", "FAILED", "CANCELLED", "REFUNDED"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "menu_item_id": { "type": "integer" },
                "quantity": { "type": "integer" },
                "price_at_time": { "type": "number", "format": "float" }
              }
            }
          }
        }
      },
      "OrderItemInput": {
        "type": "object",
        "required": ["menu_item_id", "quantity"],
        "properties": {
          "menu_item_id": { "type": "integer" },
          "quantity": { "type": "integer" }
        }
      },
      "OrderInput": {
        "type": "object",
        "required": ["user_id", "restaurant_id", "items"],
        "properties": {
          "user_id": { "type": "integer" },
          "restaurant_id": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItemInput" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/order/orders/": {
      "get": {
        "summary": "Dapatkan Semua Order",
        "tags": ["Order Service"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Daftar order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Buat Order Baru",
        "description": "Hanya Admin (berdasarkan middleware adminOnlyForMutations). Endpoint ini akan memicu alur pembayaran penuh.",
        "tags": ["Order Service"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order berhasil dibuat dan dibayar",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "400": { "description": "Pembayaran gagal (misal: saldo tidak cukup)" }
        }
      }
    },
    "/api/order/orders/{id}": {
      "get": {
        "summary": "Dapatkan Order berdasarkan ID",
        "tags": ["Order Service"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Detail order",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Perbarui Status Order (Demo)",
        "description": "Hanya Admin. Mengizinkan perubahan status order secara manual.",
        "tags": ["Order Service"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "string", "enum": ["PENDING", "PAID", "FAILED", "CANCELLED", "REFUNDED"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status order diperbarui" }
        }
      },
      "delete": {
        "summary": "Hapus Order (Demo)",
        "description": "Hanya Admin. Hanya order yang belum dibayar yang bisa dihapus.",
        "tags": ["Order Service"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Order dihapus" }
        }
      }
    },
...
  }
}